requirements.txt
mcp
pyexiftool
